feat: Add image support for blog posts and products
This change introduces the ability to associate images with blog posts and product items. It updates the `BlogPost` and `ProductItem` types to include an optional `image` field and enhances the `Blog` component to use this field for rendering post images. Additionally, product images and background video ID have been updated.